Hide stuff in subset_featurizer.py
 - rename module to `subset`
 - do not import into `msmbuilder.featurizer` namespace
