+++ msmbuilder/featurizer/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import absolute_import
 
-from .featurizer import *
-from .subset_featurizer import *
 from .feature_union import FeatureUnion
+from .featurizer import *
+from .indices import get_atompair_indices
