+++ msmbuilder/featurizer/subset.py
@@ -1,104 +1,7 @@
-import itertools
-import numpy as np
 import mdtraj as md
+import numpy as np
+
 from . import Featurizer, TrajFeatureUnion
-
-ATOM_NAMES = ["N", "CA", "CB", "C", "O", "H"]
-
-
-def get_atompair_indices(reference_traj, keep_atoms=ATOM_NAMES, exclude_atoms=None, reject_bonded=True):
-    """Get a list of acceptable atom pairs.
-
-    Parameters
-    ----------
-    reference_traj : mdtraj.Trajectory
-        Trajectory to grab atom pairs from
-    keep_atoms : np.ndarray, dtype=string, optional
-        Select only these atom names
-    exclude_atoms : np.ndarray, dtype=string, optional
-        Exclude these atom names
-    reject_bonded : bool, default=True
-        If True, exclude bonded atompairs.
-
-    Returns
-    -------
-    atom_indices : np.ndarray, dtype=int
-        The atom indices that pass your criteria
-    pair_indices : np.ndarray, dtype=int, shape=(N, 2)
-        Pairs of atom indices that pass your criteria.
-
-    Notes
-    -----
-    This function has been optimized for speed.  A naive implementation
-    can be slow (~minutes) for large proteins.
-    """
-    top, bonds = reference_traj.top.to_dataframe()
-
-    if keep_atoms is not None:
-        atom_indices = top[top.name.isin(keep_atoms) == True].index.values
-
-    if exclude_atoms is not None:
-        atom_indices = top[top.name.isin(exclude_atoms) == False].index.values
-
-    pair_indices = np.array(list(itertools.combinations(atom_indices, 2)))
-
-    if reject_bonded:
-        a_list = bonds.min(1)
-        b_list = bonds.max(1)
-
-        n = atom_indices.max() + 1
-
-        bond_hashes = a_list + b_list * n
-        pair_hashes = pair_indices[:, 0] + pair_indices[:,1] * n
-
-        not_bonds = ~np.in1d(pair_hashes, bond_hashes)
-
-        pair_indices = np.array([(a, b) for k, (a, b) in enumerate(pair_indices) if not_bonds[k]])
-
-
-    return atom_indices, pair_indices
-
-
-def _lookup_pairs_subset(all_pair_indices, subset_pair_indices, n_choose=None):
-    """Convert pairs of atom indices into a list of indices
-
-    Parameters
-    ----------
-    all_pair_indices : np.ndarray, dtype='int', shape=(N, 2)
-        All allowed pairs of atom indices
-    subset_pair_indices : np.ndarray, dtype=int, shape=(n, 2)
-        A select subset of the atom pairs
-    n_choose : int, default=None
-        if not None, return at most this many indices
-
-    Returns
-    -------
-    subset : np.ndarray, dtype=int, shape=(n)
-        A numpy array with the integer indices that map subset_pair_indices
-        onto all_pair_indices.  That is, subset[k] indices the value of
-        all_pair_indices that matches subset_pair_indices[k]
-
-
-    Notes
-    -----
-    This function is mostly useful when you have two lists of atom_pair
-    indices and you want to find "indices" mapping the smaller to the larger
-    list.  This could occur when you are looking at different atom_pair
-    featurizers.
-
-    """
-
-
-    n = all_pair_indices.max()
-
-    all_keys = all_pair_indices[:, 0] + n * all_pair_indices[:, 1]
-    optimal_keys = subset_pair_indices[:, 0] + n * subset_pair_indices[:, 1]
-    subset = np.where(np.in1d(all_keys, optimal_keys))[0]
-    if n_choose is not None:
-        subset[0:min(len(subset), n_choose)] = subset[0:min(len(subset), n_choose)]
-
-    return subset
-
 
 
 class BaseSubsetFeaturizer(Featurizer):
